+++ CHANGELOG.md
@@ -7,7 +7,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/),
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
-## [1.11.5] - Unreleased
+## [1.12.0] - Unreleased
+### Added
+- New plugin `katex` [#260].
+
 ### Changed
 - In the build mode, in order to prevent some timers to keep the process alive indefinitely and after waiting 10 seconds, exit from the Deno process with a `Deno.exit(0)`.
 
@@ -1780,8 +1783,9 @@ The first version.
 [#250]: https://github.com/lumeland/lume/issues/250
 [#255]: https://github.com/lumeland/lume/issues/255
 [#259]: https://github.com/lumeland/lume/issues/259
+[#260]: https://github.com/lumeland/lume/issues/260
 
-[1.11.5]: https://github.com/lumeland/lume/compare/v1.11.4...HEAD
+[1.12.0]: https://github.com/lumeland/lume/compare/v1.11.4...HEAD
 [1.11.4]: https://github.com/lumeland/lume/compare/v1.11.3...v1.11.4
 [1.11.3]: https://github.com/lumeland/lume/compare/v1.11.2...v1.11.3
 [1.11.2]: https://github.com/lumeland/lume/compare/v1.11.1...v1.11.2